+++ src/main/java/solutions/taulien/ribac/server/group/GroupRepository.java
@@ -1,9 +1,11 @@
 package solutions.taulien.ribac.server.group;
 
 import com.google.inject.Inject;
+import io.reactivex.Completable;
 import io.reactivex.Single;
 import solutions.taulien.ribac.server.DbHelper;
 import solutions.taulien.ribac.server.error.DuplicateCreateError;
+import solutions.taulien.ribac.server.error.ResourceNotFoundError;
 import solutions.taulien.ribac.server.gen.jooq.tables.records.DbGroup;
 import solutions.taulien.ribac.server.util.RepositoryHelper;
 
@@ -53,5 +55,26 @@ public class GroupRepository {
                                .fetch()
                                .into(DbGroup.class)
                    );
+    }
+
+
+
+    public Completable deleteGroup(String name, String requestId) {
+        return this.dbHelper
+                   .execute(
+                       requestId,
+                       db -> db.deleteFrom(GROUP)
+                               .where(GROUP.NAME.eq(name))
+                               .execute()
+                   )
+                   .map(numberOfDeletedRecords -> {
+                            if (numberOfDeletedRecords == 0) {
+                                throw new ResourceNotFoundError("A Group with the name '" + name + "' does not exist");
+                            }
+
+                            return numberOfDeletedRecords;
+                        }
+                   )
+                   .ignoreElement();
     }
 }